Readme improvements.
Improve the flow of the introduction.
Move the installation of the KTX software to it's own section.
Remove the instruction to run `npm run dist`, because it isn't needed.
Clarify and condense the note about `--no-experimental-fetch`.
Add in the `--no-experimental-fetch` option to all of the commands so that people can just copy and paste them with minimal editing.
Add a note warning the user that KTX compressed GLB files can't be imported by Blender.
Fix grammar and spelling.